requirements.txt
babel>=2.8.0,<2.8.99
channels>=2.4.0,<2.4.99
django-environ>=0.4.5,<0.4.99
djangorestframework>=3.11.0,<3.11.99
gunicorn>=20.0.4,<20.0.99
mysqlclient>=1.4.6,<1.4.99
python-dateutil>=2.8.1,<2.8.99
requests>=2.23.0,<2.23.99
requests-oauthlib>=1.3.0,<1.3.99
Django>=3.0.5,<3.0.99
